+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## v3.0.0
+
+- added `sort-json.settings.sortType`
+- Now can able to sort by `Key, Key Length, Value, Value Length`. Default is `Key`
+
 ## v2.0.0
 
 - added `sort-json.settings.isCaseSensitive` - to sort with case sensitive